# open_api_docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- {}
paths:
...
  /api/text_content/{note}/title:
    patch:
      tags:
      - text-content-controller
      operationId: updateNoteTitle
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdateTitleDTO"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NoteRealm"
  /api/text_content/{note}/details:
    patch:
      tags:
      - text-content-controller
      operationId: updateNoteDetails
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdateDetailsDTO"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NoteRealm"
...
  /api/wikidata/search/{search}:
    get:
      tags:
      - wikidata-controller
      operationId: searchWikidata
      parameters:
      - name: search
        in: path
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WikidataSearchEntity"
  /api/wikidata/entity-data/{wikidataId}:
    get:
      tags:
      - wikidata-controller
      operationId: fetchWikidataEntityDataByID
      parameters:
      - name: wikidataId
        in: path
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WikidataEntityData"
...
components:
  schemas:
    Circle:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          maxLength: 255
          minLength: 1
      required:
      - id
      - name
...
    Note:
      type: object
      properties:
        noteTopology:
          $ref: "#/components/schemas/NoteTopology"
        details:
          type: string
        parentId:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          readOnly: true
        wikidataId:
          type: string
      required:
      - createdAt
      - id
      - noteTopology
      - updatedAt
    NoteTopology:
      type: object
      properties:
        id:
          type: integer
          format: int32
        titleOrPredicate:
          type: string
        shortDetails:
          type: string
        linkType:
          type: string
          enum:
          - no link
          - related to
          - a specialization of
          - an application of
          - an instance of
          - a part of
          - tagged by
          - an attribute of
          - the opposite of
          - author of
          - using
          - an example of
          - before
          - similar to
          - confused with
        objectNoteTopology:
          $ref: "#/components/schemas/NoteTopology"
        parentOrSubjectNoteTopology:
          $ref: "#/components/schemas/NoteTopology"
        title:
          type: string
      required:
      - id
      - titleOrPredicate
    Notebook:
      type: object
      properties:
        id:
          type: integer
          format: int32
        certifiable:
          type: boolean
        notebookSettings:
          $ref: "#/components/schemas/NotebookSettings"
        creatorId:
          type: string
        title:
          type: string
        circle:
          $ref: "#/components/schemas/Circle"
        headNoteId:
          type: integer
          format: int32
        shortDetails:
          type: string
        updated_at:
          type: string
          format: date-time
      required:
      - headNoteId
      - id
      - notebookSettings
      - title
      - updated_at
    NotebookSettings:
      type: object
      properties:
        skipMemoryTrackingEntirely:
          type: boolean
        numberOfQuestionsInAssessment:
          type: integer
          format: int32
        certificateExpiry:
          type: string
          default: 1y
          example: 1y 6m
...
    NoteRealm:
      type: object
      properties:
        id:
          type: integer
          format: int32
        note:
          $ref: "#/components/schemas/Note"
        fromBazaar:
          type: boolean
        children:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        inboundReferences:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        notebook:
          $ref: "#/components/schemas/Notebook"
      required:
      - id
      - note
...
    NoteUpdateTitleDTO:
      type: object
      properties:
        newTitle:
          type: string
          maxLength: 150
          minLength: 1
      required:
      - newTitle
    NoteUpdateDetailsDTO:
      type: object
      properties:
        details:
          type: string
...
    WikidataSearchEntity:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        description:
          type: string
      required:
      - label
    WikidataEntityData:
      type: object
      properties:
        WikidataTitleInEnglish:
          type: string
        WikipediaEnglishUrl:
          type: string
      required:
      - WikidataTitleInEnglish
      - WikipediaEnglishUrl
...